pyproject.toml
[project]
name = "llm-langgraph"
version = "0.1.0"
description = ""
authors = [
    {name = "dsthiago",email = "thiagosd.2.18@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.9,<=3.12.4"
dependencies = [
    "langgraph (>=0.2.73,<0.3.0)",
    "langchain (>=0.3.18,<0.4.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "langsmith (>=0.3.8,<0.4.0)",
    "typing-extensions (>=4.12.2,<5.0.0)",
    "openai (>=1.63.0,<2.0.0)",
    "langchain-community (>=0.3.17,<0.4.0)",
    "langchain-core (>=0.3.35,<0.4.0)",
    "langchain-deepseek (>=0.1.2,<0.2.0)",
    "langchain-anthropic (>=0.3.7,<0.4.0)",
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "langchain-openai (>=0.3.6,<0.4.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "langchain-chroma (>=0.2.2,<0.3.0)",
    "docx2txt (>=0.8,<0.9)",
    "rich (>=13.9.4,<14.0.0)",
    "isort (>=6.0.0,<7.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
